requirements.txt
asgiref==3.3.1
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.7  
django-crispy-forms==1.11.1
django-filter==2.4.0
django-heroku==0.3.1
django-inline-svg==0.1.1
django-mysql==3.11.1
django-shortuuidfield==0.1.3
django-tables2==2.3.4
django-taggit==1.3.0
gunicorn==20.0.4
olefile==0.46
Pillow==8.1.1 
psycopg2==2.8.6
pytz==2021.1 
shortuuid==1.0.1
six==1.15.0
sqlparse==5.2.0
whitenoise==5.2.0
